src/requirements.txt
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Cython==0.29.24
decorator==5.1.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.27
mysqlclient==2.1.0
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
protobuf==3.19.1
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
retry==0.9.2
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.28
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
dash-bootstrap-components>=0.10.1